Backend/models/crisis_prediction/vertex_docker/requirements.txt
fastapi
uvicorn
joblib
xgboost
numpy
google-cloud-storage
